% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R, R/choropleth.R
\name{gg_choropleth}
\alias{gg_choropleth}
\alias{gg_bubbles}
\title{Ggplot choropleths by numerical variable}
\usage{
gg_bubbles(
  data = NULL,
  map_name = NULL,
  var_size = NULL,
  var_color = NULL,
  opts = NULL
)

gg_choropleth(
  data = NULL,
  map_name = NULL,
  var = NULL,
  opts = NULL,
  filter = NULL
)
}
\arguments{
\item{data}{A data.frame}
}
\value{
ggplot viz

ggplot viz
}
\description{
Ggplot choropleths by numerical variable

Ggplot choropleths by numerical variable
}
\section{ctypes}{
 Gcd-Num

 Gcd-Num
}

\examples{
gg_bubbles()
gg_choropleth_GcdNum(sample_data("Gcd-Num", nrow = 10))
}
